backup/20250919_042450/requirements.txt
crewai>=0.28.8
openai>=1.13.3
PyGithub>=2.1.1
requests>=2.31.0
jira>=3.5.2
python-dotenv>=1.0.0